Adding test to check that CLServer can handle big number of clients. Some code was added to make that true.

+++ trunk/FreeSCADA2/CLServer.Tests/ServerStarter.cs
@@ -24,17 +24,33 @@ namespace CLServer.Tests
 			process.StartInfo.Arguments = string.Format("--project-file=\"{0}\" --port=8082", project);
 			process.StartInfo.CreateNoWindow = true;
 			process.StartInfo.UseShellExecute = false;
+			process.StartInfo.RedirectStandardOutput = true;
+			process.StartInfo.RedirectStandardInput = true;
+			process.StartInfo.RedirectStandardError = true;
 
-			System.Threading.Thread.Sleep(1000); //Wait for server to load up
+			if (process.Start())
+			{
+				string line;
+				while ((line = process.StandardOutput.ReadLine()) != null)
+				{
+					if (line == "Press <ENTER> to terminate")
+						return process.HasExited == false;
+				}
+				if (process.HasExited)
+					return false;
+			}
 
-			return process.Start();
+			return false;
 		}
 
 		public void Stop()
 		{
 			if (process != null && process.HasExited == false)
 			{
-				process.Kill();
+				process.StandardInput.WriteLine("");
+				System.Threading.Thread.Sleep(500);
+				if(process.HasExited == false)
+					process.Kill();
 				process.Dispose();
 			}
 		}
